#068 - User ride history MA

+++ voom-app/app/src/main/java/inc/visor/voom/app/shared/model/VehicleType.java
@@ -1,14 +1,14 @@
 package inc.visor.voom.app.shared.model;
 
 public class VehicleType {
-    public Long id;
+    public int id;
     public String type;
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
